requirements.txt
pydantic==2.7.1
Pint==0.23
PyYAML==6.0.1